refactor: move resizing methods to pane resizer
